Clean up main page - round 1
Changed over all the links to be relative so that we can hit each
of the pages.
Converted the pages to end in .html so they work in
local development with no effort.
Pulled down the images from mainsite and placed them in the uploads
directory to make the site more self contained.