Cleaned up the Donate button in the header
Moved the CSS out into a separate file that will be the start of the
CSS clean up. Removed the use of most of the Javascript files that didn't
appear to be providing any value not that the site has been simplified.
Got the Donate button working on all the pages so that it directs
donate.html onclick