Change to external email vendor
The mail function with GoDaddy was problematic for sending emails to
Google. Likely because of the domains matching up and GMail considering
it as spam. To work around this we brough in SendInBlue email as it is
free for upto 300 emails per day.
Initial tesing has this working. Now to commit and confirm.