Change to external email vendor (#4)
The mail function with GoDaddy was problematic for sending emails to
Google. Likely because of the domains matching up and GMail considering
it as spam. To work around this we brought in SendInBlue email as it is
free for up to 300 emails per day.
Initial testing has this working. Now to commit and confirm.