Add updated news content (#8)
To get in the new content we needed to change the layout. This
allows us to have a single image in a row where we could not do
this with the way the alternating was happening but probably could
have by placing in the div for the single image rows.